# docs/swagger.yaml
basePath: /
definitions:
...
  views.Swagger_User_Login_Post:
    properties:
      email:
        example: String
        type: string
      password:
        example: String
        type: string
    type: object
  views.Swagger_User_Register_Post:
    properties:
      age:
        example: 0
        type: integer
      email:
        example: String
        type: string
      password:
        example: String
        type: string
      username:
        example: String
        type: string
    type: object
  views.Swagger_User_Update_Put:
    properties:
      email:
        example: String
        type: string
      username:
        example: String
        type: string
    type: object
host: golang-final-project-production.up.railway.app
info:
  contact:
    email: jakaprima123@gmail.com
    name: Jaka Prima Maulana
  description: final project golang jaka
  termsOfService: http://swagger.io/terms/
  title: My Grams APP
  version: v2.0
paths:
...
  /users:
    delete:
      consumes:
      - application/json
      description: Delete Data Account
      parameters:
      - description: 'Token Barier example: ''Bearer 12355f32r'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Delete Data Account
      tags:
      - USER
    put:
      consumes:
      - application/json
      description: Update Data Account
      parameters:
      - description: Update Data User
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/views.Swagger_User_Update_Put'
      - description: 'Token Barier example: ''Bearer 12355f32r'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: query
        name: userid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Update Data Account
      tags:
      - USER
  /users/login:
    post:
      consumes:
      - application/json
      description: Login Account
      parameters:
      - description: Login User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/views.Swagger_User_Login_Post'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Login Account
      tags:
      - USER
  /users/register:
    post:
      consumes:
      - application/json
      description: Register User
      parameters:
      - description: Register User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/views.Swagger_User_Register_Post'
      produces:
      - application/json
      responses:
        "201":
          description: success
          schema:
            type: string
      summary: Creates User account
      tags:
      - USER
swagger: "2.0"